pyproject.toml
[tool.poetry]
name = "spectrum-news-agent"
version = "0.1.0"
description = "An AI-powered news analysis agent that collects, summarizes, and rates news articles across the political bias spectrum."
authors = ["Divyansh Tripathi"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
openai = "*"
omegaconf = "*"
newspaper4k = "*"
nltk = "*"
lxml_html_clean = "*"
streamlit = "*"

[tool.poetry.group.dev.dependencies]
ruff = "*"
mypy = "*"
pytest = "*"
pytest-cov = "*"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "T20"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
fail_under = 90
show_missing = true
skip_empty = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
